Change rna-transcription to use Swift Package
