+++ exercises/rna-transcription/Tests/LinuxMain.swift
@@ -1,0 +1,6 @@
+import XCTest
+@testable import rnaTranscriptionTests
+
+XCTMain([
+    testCase(rnaTranscriptionTests.allTests),
+    ])
